resto de verificaciones para que cada accion de user corresponda a su player. falta que admin pueda hacer de todo

+++ app/Http/Controllers/ShotController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Shot;
 use App\Models\Player;
+use Illuminate\Support\Facades\Auth;
 
 
 class ShotController extends Controller
@@ -38,36 +39,48 @@ class ShotController extends Controller
      */
     public function store(Request $request, Player $player)
     {
-        //executem la partida amb jugador, daus i resultat
-        $player = $player->id;
-        $dice1 = rand(1,6);
-        $dice2 = rand(1,6);
+        //comprovar que el player pertany al user
+        if($player->user_id != Auth::user()->id){
+            
+            $playeruser = Player::where('user_id', Auth::user()->id)->first();
+            return response()->json(['message'=>'Aquest jugador no et pertany. El teu jugador es:',compact('playeruser')]);
+            
 
-        $result = (($dice1 + $dice2) === 7) ? true : false;    
+        }else{
+
+        //executem la partida amb jugador, daus i resultat
+            $player = $player->id;
+            $dice1 = rand(1,6);
+            $dice2 = rand(1,6);
+
+            $result = (($dice1 + $dice2) === 7) ? true : false;    
 
         //guardem la partida
-        $shot = new Shot();
+            $shot = new Shot();
 
-        $shot->dice1 = $dice1;
-        $shot->dice2 = $dice2;
-        $shot->result = $result;
-        $shot->total = ($dice1 + $dice2);
-        $shot->player_id = $player;
+            $shot->dice1 = $dice1;
+            $shot->dice2 = $dice2;
+            $shot->result = $result;
+            $shot->total = ($dice1 + $dice2);
+            $shot->player_id = $player;
 
-        $shot->save();
+            $shot->save();
 
         //actualitzem els estats del jugador
-        $playerup = Player::find($player);
-        if($result == true){
-            $playerup->increStats($playerup);
-        }else{
-            $playerup->decreStats($playerup);
-        }
+            $playerup = Player::find($player);
+            if($result == true){
+
+                $playerup->increStats($playerup);
+            }
+            else{
+
+                $playerup->decreStats($playerup);
+            }
         
         //retornem json
 
-        return response()->json(compact('player','dice1','dice2','result'));
-
+            return response()->json(compact('player','dice1','dice2','result'));
+        }
     }
 
     /**
@@ -78,9 +91,25 @@ class ShotController extends Controller
      */
     public function show(Player $player)
     {
-        $playershots = Shot::where('player_id',$player->id)->get();
+        if($player->user_id != Auth::user()->id){
+            
+            $playeruser = Player::where('user_id', Auth::user()->id)->first();
+            return response()->json(['message'=>'Aquest jugador no et pertany. El teu jugador es:',compact('playeruser')]);
+            
+        }
+        else{
+            $playershots = Shot::where('player_id',$player->id)->get();
 
-        return response()->json(compact('playershots'));
+            if(empty($playershots)){
+            
+                return response()->json(compact('playershots'));
+            
+            }
+            else{
+
+                return response()->json(['message' => 'No tens historial de jugades o aquest a sigut esborrat']);
+            }
+        }
     }
 
     /**
@@ -112,17 +141,28 @@ class ShotController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Shot $shot, $id)
+    public function destroy(Shot $shot, Player $player)
     {
-        //Esborrem jugador
-        $playershots = Shot::where("player_id","=","$id");
+        //Comprovar si les jugades pertanyen al jugador
+        if($player->user_id != Auth::user()->id){
+            
+            $playeruser = Player::where('user_id', Auth::user()->id)->first();
+            return response()->json(['message'=>'Aquest jugador no et pertany. El teu jugador es:',compact('playeruser')]);
+            
+        }
+        else{
+
+        //Esborrem jugades del jugador
+            $playershots = Shot::where("player_id","=","$player->id");
                         
-        $playershots->delete();
+            $playershots->delete();
 
         //Actualitzem estats del jugador
-        $player = Player::find($id);
-        $player->resetStats($player);
+            $player = Player::find($player->id);
+            $player->resetStats($player);
 
+            return response()->json(['message' => 'Partides esborrades correctament. El teu marcador s\'ha establert a 0']);
+        }
         
     }
